requirements.txt
Flask>=3.1
Pillow>=10.0
reportlab>=4.0
gunicorn>=22.0
